refactor: Implement generic applicative's combine and functor's fmap
Since both can be implemented in terms of a monad instance and reused
by most of the different applicatives and functors supported. Unless
they demand some specific customization, e.g. for performance reasons.

But for most cases, it simulates an automatic deriving fairly well.

NOTE: Because of minor type-level mismatches, types::either is not
using the generic implementation for now, but it shall do that in
the future.